requirements.txt
alabaster==0.7.12
alembic==0.9.9
altgraph==0.17
anaconda-client==1.7.2
anaconda-navigator==1.9.6
anaconda-project==0.8.2
antiorm==1.2.1
anytree==2.8.0
appdirs==1.4.3
asgiref==3.2.5
asn1crypto==0.24.0
astroid==2.1.0
astropy==3.1
atomicwrites==1.2.1
attrs==18.2.0
Babel==2.6.0
backcall==0.1.0
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
bcrypt==3.1.7
beautifulsoup4==4.6.3
bitarray==0.8.3
bkcharts==0.2
blaze==0.11.3
bleach==3.0.2
blinker==1.4
blis==0.4.1
bokeh==1.0.2
boto==2.49.0
bottle==0.12.18
bottle-websocket==0.2.9
Bottleneck==1.2.1
bullet==2.1.0
cefpython3==66.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
checksum==1.0.4
click==6.7
cliff==3.1.0
cloudpickle==0.6.1
clyent==1.2.2
cmd2==0.8.9
colorama==0.4.1
comtypes==1.1.7
conda==4.7.6
conda-build==3.17.6
conda-package-handling==1.3.11
conda-verify==3.1.1
configobj==5.0.6
configparser==5.0.0
contextlib2==0.5.5
cryptography==2.4.2
cycler==0.10.0
cymem==2.0.3
Cython==0.29.2
cytoolz==0.9.0.1
dask==1.0.0
datashape==0.5.4
db==0.1.1
db-sqlite3==0.0.1
decorator==4.3.0
defusedxml==0.5.0
distlib==0.3.0
distributed==1.25.1
Django==3.0
django-allauth==0.42.0
django-crispy-forms==1.9.1
django-filter==2.3.0
django-multiselectfield==0.1.12
django-partial-date==1.3.1
django-taggit==1.3.0
djangorestframework==3.11.0
docutils==0.14
Eel==0.12.4
entrypoints==0.2.3
et-xmlfile==1.0.1
extract-zip==1.0.0
fastcache==1.0.2
filelock==3.0.10
Flask==1.0.2
Flask-Admin==1.5.6
Flask-Cors==3.0.7
Flask-Dance==0.14.0
Flask-DebugToolbar==0.10.1
Flask-GoogleCharts==0.1.0
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-MySQLdb==0.2.0
Flask-OAuth==0.12
Flask-OAuthlib==0.9.4
Flask-SQLAlchemy==2.3.2
Flask-User==1.0.2.2
Flask-WTF==0.14.2
future==0.17.1
geocoder==1.38.1
gevent==1.3.7
gevent-websocket==0.10.1
glob2==0.6
gnupg==2.3.1
greenlet==0.4.15
gunicorn==20.0.4
h5py==2.8.0
heapdict==1.0.0
html5lib==1.0.1
httplib2==0.11.3
humanfriendly==8.2
idna==2.6
imageio==2.4.1
imagesize==1.1.0
importlib-metadata==1.5.0
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.4
itsdangerous==0.24
jdcal==1.4
jedi==0.13.2
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyterlab==0.35.3
jupyterlab-server==0.2.0
keyring==17.0.0
Kivy-Garden==0.1.4
kiwisolver==1.0.1
lazy==1.3
lazy-object-proxy==1.3.1
libarchive-c==2.8
llvmlite==0.26.0
locket==0.2.0
lxml==4.2.5
Mako==1.0.7
MarkupSafe==1.0
mathparse==0.1.2
matplotlib==3.0.2
mccabe==0.6.1
menuinst==1.4.14
mistune==0.8.4
mkl-fft==1.0.6
mkl-random==1.0.2
more-itertools==4.3.0
mpmath==1.1.0
msgpack==0.5.6
multipledispatch==0.6.0
murmurhash==1.0.2
mysql-connector==2.2.9
mysqlclient==1.4.6
navigator-updater==0.2.1
nbconvert==5.4.0
nbformat==4.4.0
netifaces==0.10.9
networkx==2.2
nltk==3.4
nose==1.3.7
notebook==5.7.4
numba==0.41.0
numexpr==2.6.8
numpy==1.15.4
numpydoc==0.8.0
oauth2==1.9.0.post1
oauthlib==2.0.7
odo==0.5.1
olefile==0.46
openpyxl==2.5.12
os-sys-php==2019.10.13
packaging==18.0
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
partd==0.3.9
passlib==1.7.2
path.py==11.5.0
pathlib2==2.3.3
patsy==0.5.1
paytm==0.1.8
pbkdf2==1.3
pbr==5.4.5
pefile==2019.4.18
pep8==1.7.1
pexpect==4.8.0
pickle-mixin==1.0.2
pickleshare==0.7.5
Pillow==5.3.0
Pint==0.12
pkginfo==1.4.2
plac==0.9.6
pluggy==0.13.1
ply==3.11
preshed==3.0.2
prettytable==0.7.2
progress==1.5
progressbar==2.5
prometheus-client==0.5.0
prompt-toolkit==2.0.7
property-manager==3.0
psutil==5.4.8
ptyprocess==0.6.0
py==1.7.0
pycodestyle==2.4.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycryptodome==3.9.8
pycurl==7.43.0.2
pydotplus==2.0.2
pyflakes==2.0.0
Pygments==2.3.1
pygubu==0.10
PyInstaller==3.6
pylint==2.2.2
PyMySQL==0.9.3
pyodbc==4.0.25
pyOpenSSL==18.0.0
pyotp==2.3.0
pyparsing==2.3.0
pyperclip==1.8.0
pypiwin32==223
pyreadline==2.1
PySocks==1.6.8
pyspeedtest==1.2.7
pytest==4.0.2
pytest-arraydiff==0.3
pytest-astropy==0.5.0
pytest-doctestplus==0.2.0
pytest-openfiles==0.3.1
pytest-remotedata==0.3.1
python-dateutil==2.7.2
python-editor==1.0.3
python-secrets==20.4.2106
python3-openid==3.2.0
pythonGUI==0.0.2
pytz==2018.7
pyvalid==0.9.2
PyWavelets==1.0.1
pywin32==223
pywin32-ctypes==0.2.0
pywinpty==0.5.5
PyYAML==3.13
pyzmq==17.1.2
QtAwesome==0.5.3
qtconsole==4.4.3
QtPy==1.5.2
ratelim==0.1.6
razorpay==1.2.0
requests==2.24.0
requests-oauthlib==0.8.0
rope==0.11.0
ruamel-yaml==0.15.46
scikit-image==0.14.1
scikit-learn==0.20.1
scipy==1.4.1
seaborn==0.10.1
selenium==3.141.0
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
snowballstemmer==1.2.1
sortedcollections==1.0.1
sortedcontainers==2.1.0
spacy==2.2.0
Sphinx==1.8.2
sphinxcontrib-websupport==1.1.0
spyder==3.3.2
spyder-kernels==0.3.0
SQLAlchemy==1.3.17
SQLAlchemy-Utils==0.33.2
sqlparse==0.3.0
srsly==1.0.2
statsmodels==0.9.0
stevedore==1.32.0
stripe==2.48.0
sympy==1.3
tables==3.4.4
tblib==1.3.2
terminado==0.8.1
testpath==0.4.2
text-editor==1.0.5
thinc==7.1.1
tinydb==4.1.1
toml==0.10.1
toolz==0.9.0
tornado==5.1.1
tox==3.15.1
tqdm==4.28.1
traitlets==4.3.2
tuspy==0.2.4
unicodecsv==0.14.1
urllib3==1.22
URLObject==2.4.3
verboselogs==1.7
virtualenv==20.0.24
wasabi==0.6.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
whichcraft==0.6.1
whitenoise==5.1.0
widgetsnbextension==3.4.2
wifi==0.3.8
win-inet-pton==1.0.1
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.10.11
WTForms==2.1
xgboost==1.0.2
xkcdpass==1.17.3
xlrd==1.2.0
XlsxWriter==1.1.2
xlwings==0.15.1
xlwt==1.3.0
yamlreader==3.0.4
zict==0.1.3
zipp==3.0.0
